Move JsonElementExtensions to Extensions folder, add blank lines per style
- Move OAuth2/Infrastructure/JsonElementExtensions.cs -> OAuth2/Extensions/JsonElementExtensions.cs
- Change namespace from OAuth2.Infrastructure to OAuth2.Extensions
- Add using OAuth2.Extensions to all consuming files
- Add blank lines after return/closing braces per code style
