+++ OAuth2/Client/Impl/LoginCidadaoClient.cs
@@ -7,6 +7,7 @@ using System.Threading.Tasks;
 using OAuth2.Client;
 using OAuth2.Configuration;
 using OAuth2.Infrastructure;
+using OAuth2.Extensions;
 using OAuth2.Models;
 using RestSharp;
