+++ OAuth2/Client/Impl/OdnoklassnikiClient.cs
@@ -3,6 +3,7 @@ using System.Linq;
 using System.Text.Json;
 using OAuth2.Configuration;
 using OAuth2.Infrastructure;
+using OAuth2.Extensions;
 using OAuth2.Models;
 using RestSharp;
