+++ OAuth2/Client/Impl/VkClient.cs
@@ -1,6 +1,7 @@
 using System.Text.Json;
 using OAuth2.Configuration;
 using OAuth2.Infrastructure;
+using OAuth2.Extensions;
 using OAuth2.Models;
 using RestSharp;
