+++ OAuth2/Client/OAuth2Client.cs
@@ -6,6 +6,7 @@ using System.Threading.Tasks;
 using System.Web;
 using OAuth2.Configuration;
 using OAuth2.Infrastructure;
+using OAuth2.Extensions;
 using OAuth2.Models;
 using RestSharp;
 using RestSharp.Authenticators;
